src/requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
autocommand==2.2.2
backports.tarfile==1.2.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docstring_parser==0.16
dptr==1.0.0
einops==0.8.0
einshape==1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.54.1
fsspec==2024.9.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
grpcio==1.66.2
huggingface-hub==0.25.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iopath==0.1.10
ipython==8.27.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jaxtyping==0.2.34
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
more-itertools==10.3.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp310-cp310-linux_x86_64.whl#sha256=120568f3fd675f59e4cb6de79a5c193b4067d90878d8bab2040fda8e3a2df6fa
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.2.2
plyfile==1.1
portalocker==2.10.1
prompt_toolkit==3.0.48
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytorch3d==0.7.7
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
rich==13.9.1
safetensors==0.4.5
scipy==1.14.1
shtab==1.7.1
-e git+https://gitlab.inria.fr/bkerbl/simple-knn@86710c2d4b46680c02301765dd79e465819c8f19#egg=simple_knn
six==1.16.0
stack-data==0.6.3
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tokenizers==0.20.0
tomli==2.0.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
triton==3.0.0
typeguard==2.13.3
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tyro==0.8.11
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.19.2
